+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for PommaLabs.Thrower #
 
+### v4.0.5 (2016-12-??) ###
+
+* EnumerationAttribute doc comments contained a few mistakes.
+* EmailAddressAttribute and PhoneNumberAttribute now validate successfully null values.
+* Fixed and enriched unit tests for object validation.
+
 ### v4.0.4 (2016-12-17) ###
 
 * All code marked as Obsolete in v3 has now been completely removed.